+++ apps/meteor/client/omnichannel/businessHours/useRemoveBusinessHour.tsx
@@ -1,4 +1,4 @@
-import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
+import { useEffectEvent } from '@rocket.chat/fuselage-hooks';
 import { useSetModal, useToastMessageDispatch, useMethod } from '@rocket.chat/ui-contexts';
 import { useQueryClient } from '@tanstack/react-query';
 import { useTranslation } from 'react-i18next';
@@ -12,7 +12,7 @@ export const useRemoveBusinessHour = () => {
 	const removeBusinessHour = useMethod('livechat:removeBusinessHour');
 	const queryClient = useQueryClient();
 
-	const handleRemove = useMutableCallback((_id, type) => {
+	const handleRemove = useEffectEvent((_id: string, type: string) => {
 		const onDeleteBusinessHour = async () => {
 			try {
 				await removeBusinessHour(_id, type);