+++ apps/meteor/client/sidebar/sections/OmnichannelSection.tsx
@@ -1,5 +1,5 @@
 import { Sidebar, SidebarDivider, SidebarSection } from '@rocket.chat/fuselage';
-import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
+import { useEffectEvent } from '@rocket.chat/fuselage-hooks';
 import { FeaturePreview, FeaturePreviewOff, FeaturePreviewOn } from '@rocket.chat/ui-client';
 import { useLayout, useRoute, usePermission } from '@rocket.chat/ui-contexts';
 import { memo } from 'react';
@@ -23,7 +23,7 @@ const OmnichannelSection = () => {
 	const queueListRoute = useRoute('livechat-queue');
 	const isWorkspaceOverMacLimit = useIsOverMacLimit();
 
-	const handleRoute = useMutableCallback((route) => {
+	const handleRoute = useEffectEvent((route: string) => {
 		sidebar.toggle();
 
 		switch (route) {