+++ apps/meteor/client/views/e2e/EnterE2EPasswordModal.tsx
@@ -1,6 +1,7 @@
 import { Box, PasswordInput, Field, FieldGroup, FieldRow, FieldError } from '@rocket.chat/fuselage';
-import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
-import type { ChangeEvent, ReactElement } from 'react';
+import { useEffectEvent } from '@rocket.chat/fuselage-hooks';
+import DOMPurify from 'dompurify';
+import type { ChangeEvent, FormEvent, ReactElement } from 'react';
 import { useState, useCallback } from 'react';
 import { useTranslation } from 'react-i18next';
 
@@ -27,7 +28,7 @@ const EnterE2EPasswordModal = ({
 		[setPassword],
 	);
 
-	const handleConfirm = useMutableCallback((e): void => {
+	const handleConfirm = useEffectEvent((e: FormEvent): void => {
 		e.preventDefault();
 		if (password === '') {
 			setPasswordError(t('Invalid_pass'));
@@ -48,7 +49,7 @@ const EnterE2EPasswordModal = ({
 			onClose={onClose}
 			onCancel={onCancel}
 		>
-			<Box dangerouslySetInnerHTML={{ __html: t('E2E_password_request_text') }} />
+			<Box dangerouslySetInnerHTML={{ __html: DOMPurify.sanitize(t('E2E_password_request_text')) }} />
 			<FieldGroup mbs={24} w='full'>
 				<Field>
 					<FieldRow>