+++ apps/meteor/client/views/omnichannel/agents/hooks/useRemoveAgent.tsx
@@ -1,5 +1,5 @@
 import type { ILivechatAgent } from '@rocket.chat/core-typings';
-import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
+import { useEffectEvent } from '@rocket.chat/fuselage-hooks';
 import { useSetModal, useToastMessageDispatch, useTranslation, useRouter, useEndpoint } from '@rocket.chat/ui-contexts';
 import { useQueryClient } from '@tanstack/react-query';
 
@@ -14,7 +14,7 @@ export const useRemoveAgent = (uid: ILivechatAgent['_id']) => {
 
 	const deleteAction = useEndpoint('DELETE', '/v1/livechat/users/agent/:_id', { _id: uid });
 
-	const handleDelete = useMutableCallback(() => {
+	const handleDelete = useEffectEvent(() => {
 		const onDeleteAgent = async () => {
 			try {
 				await deleteAction();