+++ apps/meteor/client/views/omnichannel/departments/DepartmentsPage.tsx
@@ -1,5 +1,5 @@
 import { Tabs, Button } from '@rocket.chat/fuselage';
-import { useMutableCallback } from '@rocket.chat/fuselage-hooks';
+import { useEffectEvent } from '@rocket.chat/fuselage-hooks';
 import { useRoute, useTranslation, useRouteParameter } from '@rocket.chat/ui-contexts';
 
 import DepartmentsTableV2 from './DepartmentsTable';
@@ -14,13 +14,17 @@ const DepartmentsPage = () => {
 	const context = useRouteParameter('context');
 	const id = useRouteParameter('id');
 
-	const handleTabClick = useMutableCallback((tab) =>
-		departmentsRoute.push({
-			context: tab,
-		}),
+	const handleTabClick = useEffectEvent((tab: undefined | 'archived') =>
+		departmentsRoute.push(
+			tab
+				? {
+						context: tab,
+					}
+				: {},
+		),
 	);
 
-	const onAddNew = useMutableCallback(() =>
+	const onAddNew = useEffectEvent(() =>
 		departmentsRoute.push({
 			context: 'new',
 		}),