+++ packages/api-client/CHANGELOG.md
@@ -1,5 +1,15 @@
 # @rocket.chat/api-client
 
+## 0.2.12
+
+### Patch Changes
+
+- <details><summary>Updated dependencies []:</summary>
+
+  - @rocket.chat/core-typings@7.2.1
+  - @rocket.chat/rest-typings@7.2.1
+  </details>
+
 ## 0.2.11
 
 ### Patch Changes